+++ src/backend/api/lambda.js
@@ -1,0 +1,4 @@
+import serverlessExpress from "@codegenie/serverless-express";
+import app from "./app.js";
+
+exports.handler = serverlessExpress({ app });