+++ src/backend/api/routes/index.js
@@ -1,0 +1,7 @@
+import filesRoutes from "./files.js";
+import foldersRoutes from "./folders.js";
+
+export default (router) => {
+  filesRoutes(router);
+  foldersRoutes(router);
+};